update api calls on each move
before, the game was just being played locally and being sent to the server. now, the
game on the server gets updated with each move, and the board gets reprinted from
the server after every move. this allows for multiplayer games and the ability to
end a multiplayer game mid-stream.